Functions for PBC and scatter matrix plots.
Function PBC() was created to calculate the Point-Biserial Correlation of all numeric columns of a DataFrame. Functions scatter_matrix() and my_scatter_matrix() were added to plot scatter matrix graphics. The former uses the Seaborn pairplot function and will be deleted in the future. The latter is prefered and plots the univariate distributions using Gaussian kernels on top of the matrix (as if they were the matrix header).